move tg_bot_page to pages, restructure templates folder
