+++ backend/django_mem_cards/urls.py
@@ -9,8 +9,6 @@ from .views import RegistrationView
 urlpatterns = [
     path('', include('homepage.urls', namespace='homepage')),
     path('pages/', include('pages.urls', namespace='pages')),
-    #path('about/', include('about.urls', namespace='about')),
-    path('tg_bot_page/', include('tg_bot_page.urls', namespace='tg_bot_page')),
     path('deck/', include('deck.urls', namespace='deck')),
     path('admin/', admin.site.urls),
     path('auth/', include('django.contrib.auth.urls')),
