+++ backend/templates/homepage/index.html
@@ -84,7 +84,7 @@
 
           <a href="{% url 'password_change' %}" class="btn btn-outline-secondary col-8">&#x270e; Сменить пароль</a>
           <hr class="opacity-0">
-          <a href="{% url 'tg_bot_page:tg_bot_page_view' %}" class="btn btn-outline-secondary col-8">
+          <a href="{% url 'pages:tg_bot_page_view' %}" class="btn btn-outline-secondary col-8">
             <img class="icon" src="{% static 'img/TG_logo.svg' %}" width=18 height=18>
              Telegram-бот
           </a>
