--- backend/tg_bot_page/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class TgBotPageConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'tg_bot_page'
